\name{rocdata}
\docType{data}
\alias{rocdata}
\title{Example data set for continuos tests results}
\description{
This data set gives results of two serological tests - continuous scale - and its
correspondig reference test classfication - 0 for negative and 1 for positive.
}
\usage{rocdata}
\format{A data.frame containing 3 variables and 148 observations.}
\source{Not informed}
\references{
Not informed
}
\keyword{datasets}